view refactoring, 新着情報の専用クラス復活
