+++ web/views/__init__.py
@@ -1,0 +1,2 @@
+from web.views.main_page import *
+from web.views.question import *
