+++ web/views/basic_info/__init__.py
@@ -1,0 +1,4 @@
+from web.views.basic_info.flow_info import *
+from web.views.basic_info.price_info import *
+from web.views.basic_info.advantage_info import *
+from web.views.basic_info.company_info import *
