+++ web/views/basic_info/advantage_info.py
@@ -1,0 +1,5 @@
+from django.views.generic import TemplateView
+
+class AdvantageInfo(TemplateView):
+    """XXXの強み"""
+    template_name = "web/info_advantage.html"
