+++ web/views/basic_info/flow_info.py
@@ -1,0 +1,5 @@
+from django.views.generic import TemplateView
+
+class FlowInfo(TemplateView):
+    """装置を売る＞掲載〜契約までの流れ"""
+    template_name = "web/info_flow.html"
