+++ web/views/case/__init__.py
@@ -1,0 +1,5 @@
+from web.views.case.case_list import *
+from web.views.case.case_detail import *
+from web.views.case.case_create import *
+from web.views.case.case_my_list import *
+from web.views.case.case_update import *
