+++ web/views/favorite/__init__.py
@@ -1,0 +1,2 @@
+from web.views.favorite.favorite import *
+from web.views.favorite.favorite_list import *
