+++ web/views/item/__init__.py
@@ -1,0 +1,8 @@
+from web.views.item.item_accept import *
+from web.views.item.item_create import *
+from web.views.item.item_detail import *
+from web.views.item.item_list import *
+from web.views.item.item_my_list import *
+from web.views.item.item_new_list import *
+from web.views.item.item_update import *
+from web.views.item.item_new_list import *
