+++ web/views/main_page.py
@@ -1,0 +1,5 @@
+from django.views.generic import TemplateView
+
+class Main(TemplateView):
+    """トップページ"""
+    template_name = "web/main.html"
