+++ web/views/search_by/__init__.py
@@ -1,0 +1,2 @@
+from web.views.search_by.area_list import *
+from web.views.search_by.item_type_list import *
