+++ web/views/search_by/area_list.py
@@ -1,0 +1,6 @@
+from django.views.generic import ListView
+from web.models import Area
+
+class AreaList(ListView):
+    model = Area
+    template_name = "web/area_list.html"
