+++ web/views/search_by/item_type_list.py
@@ -1,0 +1,6 @@
+from django.views.generic import ListView
+from web.models import ItemType
+
+class ItemTypeList(ListView):
+    model = ItemType
+    template_name = "web/itemtype_list.html"
